[add] Upd001_トークテーマ投稿の実装

+++ eternal-app/src/main/java/com/eternal/web/repository/CommentRepository.java
@@ -3,7 +3,6 @@
  */
 package com.eternal.web.repository;
 
-import java.math.BigDecimal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.eternal.web.entity.CommentEntity;
@@ -14,6 +13,6 @@ import com.eternal.web.entity.CommentEntity;
  * @author taiki0304
  */
 @Repository
-public interface CommentRepository extends JpaRepository<CommentEntity, BigDecimal> {
+public interface CommentRepository extends JpaRepository<CommentEntity, Long> {
 
 }
