+++ eternal-app/src/main/java/com/eternal/web/repository/TalkEvaluationRepository.java
@@ -3,7 +3,6 @@
  */
 package com.eternal.web.repository;
 
-import java.math.BigDecimal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.eternal.web.entity.TalkEvaluationEntity;
@@ -14,6 +13,6 @@ import com.eternal.web.entity.TalkEvaluationEntity;
  * @author taiki0304
  */
 @Repository
-public interface TalkEvaluationRepository extends JpaRepository<TalkEvaluationEntity, BigDecimal> {
+public interface TalkEvaluationRepository extends JpaRepository<TalkEvaluationEntity, Long> {
 
 }
