+++ eternal-app/src/main/java/com/eternal/web/repository/TalkThemeRepository.java
@@ -1,6 +1,5 @@
 package com.eternal.web.repository;
 
-import java.math.BigDecimal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.eternal.web.entity.TalkThemeEntity;
@@ -11,6 +10,6 @@ import com.eternal.web.entity.TalkThemeEntity;
  * @author taiki0304
  */
 @Repository
-public interface TalkThemeRepository extends JpaRepository<TalkThemeEntity, BigDecimal> {
+public interface TalkThemeRepository extends JpaRepository<TalkThemeEntity, Long> {
 
 }
