+++ eternal-app/src/main/java/com/eternal/web/repository/UserRepository.java
@@ -3,7 +3,6 @@
  */
 package com.eternal.web.repository;
 
-import java.math.BigDecimal;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -15,7 +14,7 @@ import com.eternal.web.entity.UserEntity;
  * @author taiki0304
  */
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, BigDecimal> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
     public Optional<UserEntity> findByUserName(String userName);
 }
